fix: break circular import in UTM analytics helpers
